debugging

+++ models/utils.py
@@ -172,7 +172,7 @@ def get_score_fn(sde, model, train=False, continuous=False):
         else:
           # For VE-trained models, t=0 corresponds to the highest noise level
           labels = t*(sde[1].N - 1)
-          labels = torch.round(labels).long()
+          labels = torch.round(labels.float()).long()
           score = model_fn(x, labels)
           score = divide_by_sigmas(score, labels, sde)
